Fix GCC warning
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@16950 f0d54148-0727-0410-94bb-9a71ac55c965

+++ apps/gdaldem.cpp
@@ -813,9 +813,9 @@ int GDALColorReliefFindNamedColor(const char *pszColorName, int *pnR, int *pnG, 
     {
         if (EQUAL(pszColorName, namedColors[i].name))
         {
-            *pnR = 255. * namedColors[i].r;
-            *pnG = 255. * namedColors[i].g;
-            *pnB = 255. * namedColors[i].b;
+            *pnR = (int)(255. * namedColors[i].r);
+            *pnG = (int)(255. * namedColors[i].g);
+            *pnB = (int)(255. * namedColors[i].b);
             return TRUE;
         }
     }
